scripts/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiohttp==3.9.5
    # via
    #   -r requirements.in
    #   microsoft-kiota-authentication-azure
aiosignal==1.3.1
    # via aiohttp
anyio==4.3.0
    # via httpx
attrs==23.2.0
    # via aiohttp
azure-core==1.30.1
    # via
    #   azure-identity
    #   microsoft-kiota-authentication-azure
azure-identity==1.16.0
    # via
    #   -r requirements.in
    #   msgraph-beta-sdk
    #   msgraph-sdk
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
cryptography==42.0.7
    # via
    #   azure-identity
    #   msal
    #   pyjwt
deprecated==1.2.14
    # via opentelemetry-api
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
h11==0.14.0
    # via httpcore
h2==4.1.0
    # via httpx
hpack==4.0.0
    # via h2
httpcore==1.0.5
    # via httpx
httpx[http2]==0.27.0
    # via
    #   microsoft-kiota-http
    #   msgraph-core
hyperframe==6.0.1
    # via h2
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==7.0.0
    # via opentelemetry-api
markdown-it-py==3.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
microsoft-kiota-abstractions==1.3.2
    # via
    #   microsoft-kiota-authentication-azure
    #   microsoft-kiota-http
    #   microsoft-kiota-serialization-form
    #   microsoft-kiota-serialization-json
    #   microsoft-kiota-serialization-multipart
    #   microsoft-kiota-serialization-text
    #   msgraph-beta-sdk
    #   msgraph-core
    #   msgraph-sdk
microsoft-kiota-authentication-azure==1.0.0
    # via
    #   msgraph-beta-sdk
    #   msgraph-core
    #   msgraph-sdk
microsoft-kiota-http==1.3.1
    # via
    #   msgraph-beta-sdk
    #   msgraph-core
    #   msgraph-sdk
microsoft-kiota-serialization-form==0.1.0
    # via
    #   msgraph-beta-sdk
    #   msgraph-sdk
microsoft-kiota-serialization-json==1.2.0
    # via
    #   msgraph-beta-sdk
    #   msgraph-sdk
microsoft-kiota-serialization-multipart==0.1.0
    # via
    #   msgraph-beta-sdk
    #   msgraph-sdk
microsoft-kiota-serialization-text==1.0.0
    # via
    #   msgraph-beta-sdk
    #   msgraph-sdk
msal==1.28.0
    # via
    #   azure-identity
    #   msal-extensions
msal-extensions==1.1.0
    # via azure-identity
msgraph-beta-sdk==1.4.0
    # via -r requirements.in
msgraph-core==1.0.0
    # via
    #   msgraph-beta-sdk
    #   msgraph-sdk
msgraph-sdk==1.4.0
    # via -r requirements.in
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
opentelemetry-api==1.24.0
    # via
    #   microsoft-kiota-abstractions
    #   microsoft-kiota-authentication-azure
    #   microsoft-kiota-http
    #   opentelemetry-sdk
opentelemetry-sdk==1.24.0
    # via
    #   microsoft-kiota-abstractions
    #   microsoft-kiota-authentication-azure
    #   microsoft-kiota-http
opentelemetry-semantic-conventions==0.45b0
    # via opentelemetry-sdk
packaging==24.0
    # via msal-extensions
pendulum==3.0.0
    # via
    #   microsoft-kiota-serialization-form
    #   microsoft-kiota-serialization-json
portalocker==2.8.2
    # via msal-extensions
pycparser==2.22
    # via cffi
pygments==2.18.0
    # via rich
pyjwt[crypto]==2.8.0
    # via msal
python-dateutil==2.9.0.post0
    # via
    #   microsoft-kiota-serialization-text
    #   pendulum
    #   time-machine
python-dotenv==1.0.1
    # via -r requirements.in
requests==2.31.0
    # via
    #   azure-core
    #   msal
rich==13.7.1
    # via -r requirements.in
six==1.16.0
    # via
    #   azure-core
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
std-uritemplate==0.0.57
    # via microsoft-kiota-abstractions
time-machine==2.14.1
    # via pendulum
typing-extensions==4.11.0
    # via
    #   azure-core
    #   opentelemetry-sdk
tzdata==2024.1
    # via pendulum
urllib3==2.2.1
    # via requests
wrapt==1.16.0
    # via deprecated
yarl==1.9.4
    # via aiohttp
zipp==3.18.2
    # via importlib-metadata
